pyproject.toml
[tool.poetry]
name = "permissions"
version = "0.1.0"
description = ""
authors = ["Chenge"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlmodel = "^0.0.21"
sqlalchemy-seeder = "^0.3.0"
setuptools = "^71.1.0"
python-decouple = "^3.8"
bcrypt = "^4.2.0"
pendulum = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
